+++ src/extension.ts
@@ -1,8 +1,8 @@
 import { commands, ExtensionContext } from 'vscode'
-import { executeGenerateTemplateCommand } from './templates/template-generation'
+import { executeGenerateTemplateCommand } from './templates/template-generator'
 
 export function activate (context: ExtensionContext) {
-	const disposable = commands.registerCommand('componentgenerator.generateComponent', async e => {
+	const disposable = commands.registerCommand('templateGenerator.generateTemplate', async e => {
 		await executeGenerateTemplateCommand(e && e.fsPath)
 	})